+++ CHANGELOG.md
@@ -58,6 +58,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - PNPM v5 lockfile support
+- `phylum project update --default-label`
 
 ### Fixed